Fake user email if nto provided by JIRA
In some instances, JIRA will have email visibility turned
off for the REST API. In such instances, we should use the
domain of the JIRA instance and the username (which is still
visible to us) to fake an email address. While not perfect,
this still allows us to get a good unique count of actors.